refactor(*): restructure `app` & `mvc`
- reformat all sources
