--- app/src/main/java/com/techcourse/ControllerHandlerAdapter.java
@@ -1,29 +1,0 @@
-package com.techcourse;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request, final HttpServletResponse response
-            , final Object handler) throws Exception {
-        if (!supports(handler)) {
-            throw new IllegalArgumentException("This handler is not supported: " + handler);
-        }
-
-        Controller controller = (Controller) handler;
-        JspView view = new JspView(controller.execute(request, response));
-        return new ModelAndView(view);
-    }
-}
