--- app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -1,60 +1,0 @@
-package com.techcourse;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecutionAdapter;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.Map;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class DispatcherServlet extends HttpServlet {
-
-    private static final long serialVersionUID = 1L;
-    private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
-
-    private final HandlerMappingRegistry handlerMappingRegistry = new HandlerMappingRegistry();
-    private final HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
-
-    public DispatcherServlet() {
-    }
-
-    @Override
-    public void init() {
-        try {
-            handlerMappingRegistry.registerHandlerMapping(ManualHandlerMapping.class);
-            handlerAdapterRegistry.registerHandlerAdapter(ControllerHandlerAdapter.class);
-
-            handlerMappingRegistry.registerHandlerMapping(AnnotationHandlerMapping.class, "samples");
-            handlerAdapterRegistry.registerHandlerAdapter(HandlerExecutionAdapter.class);
-        } catch (ReflectiveOperationException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    protected void service(final HttpServletRequest request, final HttpServletResponse response)
-            throws ServletException {
-        final String requestURI = request.getRequestURI();
-        log.debug("Method : {}, Request URI : {}", request.getMethod(), requestURI);
-
-        try {
-            Object handler = handlerMappingRegistry.getHandler(request);
-            HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(handler);
-
-            ModelAndView modelAndView = handlerAdapter.handle(request, response, handler);
-            View view = modelAndView.getView();
-            Map<String, Object> model = modelAndView.getModel();
-
-            view.render(model, request, response);
-        } catch (Throwable e) {
-            log.error("Exception : {}", e.getMessage(), e);
-            throw new ServletException(e.getMessage());
-        }
-    }
-}
