+++ app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,6 +1,6 @@
 package com.techcourse.controller;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.legacy.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
