+++ mvc/src/main/java/com/interface21/context/stereotype/ControllerScanner.java
@@ -1,6 +1,5 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.context.stereotype;
 
-import com.interface21.context.stereotype.Controller;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Set;
@@ -8,7 +7,8 @@ import java.util.stream.Collectors;
 import org.reflections.Reflections;
 
 public class ControllerScanner {
-    public Set<Class<?>> scan(final Object[] packages) {
+
+    public Set<Class<?>> scanControllerClasses(final Object[] packages) {
         return Arrays.stream(packages).map(Reflections::new)
                 .map(r -> r.getTypesAnnotatedWith(Controller.class))
                 .flatMap(Collection::stream)
