+++ mvc/src/main/java/com/interface21/core/servlet/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.core.servlet;
 
 public class UncheckedServletException extends RuntimeException {
