+++ mvc/src/main/java/com/interface21/web/handler/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.web.handler.adapter;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.web.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
