+++ mvc/src/main/java/com/interface21/web/view/View.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.web.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
