+++ mvc/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,1 @@
-com.interface21.web.SpringServletContainerInitializer
+com.interface21.core.SpringServletContainerInitializer
