+++ mvc/src/test/java/samples/TestController.java
@@ -1,10 +1,10 @@
 package samples;
 
 import com.interface21.context.stereotype.Controller;
+import com.interface21.web.ModelAndView;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
+import com.interface21.web.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
